Create CVE-2024-394. nse
Checks for the xz compression utility vulnerability (CVE-2024-3094) in versions 5.6.0 and 5.6.1.

This script connects to the target host and executes the 'xz --version' command to retrieve the version information of the xz compression utility. It then checks if the version is either 5.6.0 or 5.6.1. If the version matches, it indicates a vulnerability (CVE-2024-3094) in the xz utility, otherwise, it reports no vulnerability found.